+++ Automation/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v4.7.0 (TBD)
 * [Feature] Add an action for setting Gate state.
+* [Feature] Add manufactory signals for the current recipe and recipe completion.
 * [Fix] Manufactory input and output goods could be missing from the building signal export list.
 
 # v4.6.4 (June 29th, 2026)